Development

+++ src/server/api_ihrmis/database/migrations/2022_06_17_024123_create_tblemployee_cseligibilities_table.php
@@ -19,7 +19,7 @@ class CreateTblemployeeCseligibilitiesTable extends Migration
             $table->string('cse_emp_title', 150)->comment('Title or name of the Career Service/ RA1080 (Board/ Bar) Under Special Laws/ CES/ CSEE/ Barangay Eligibility/ Driver\'s License earned.');
             $table->date('cse_emp_date')->comment('Date the eligibility exam was taken.');
             $table->string('cse_emp_place', 150)->comment('Place where the eligibility exam took place.');
-            $table->float('cse_emp_rating', 5)->unsigned()->comment('Final rating achieved from the eligibility exam taken.');
+            $table->double('cse_emp_rating', 5, 2)->unsigned()->comment('Final rating achieved from the eligibility exam taken.');
             $table->string('cse_emp_license', 30)->comment('License number if earned eligibility entails a license.');
             $table->date('cse_emp_validity')->comment('Date of validity if earned eligibility entails a license.');
         });
